comment can upload

+++ server/sqlhelp/blog.sql
@@ -310,7 +310,7 @@ CREATE TABLE  IF NOT EXISTS user_comment (
  comment_target_id mediumint(8) NOT NULL COMMENT '评论内容的ID',
  comment_content varchar(255) NOT NULL COMMENT '评论内容',
  commenter_user_id mediumint(8) NOT NULL COMMENT '评论者ID',
- comment_time int(13) NOT NULL COMMENT '评论时间',
+ comment_time BIGINT(15) NOT NULL COMMENT '评论时间',
  commenter_ip varchar(15) NOT NULL COMMENT '评论时的IP地址',
  delete_flag bit NOT NULL DEFAULT 0 COMMENT '删除标志',
  PRIMARY KEY (comment_id)
@@ -327,7 +327,7 @@ CREATE TABLE  IF NOT EXISTS user_sub_comment (
  comment_target_id mediumint(8) NOT NULL COMMENT '评论内容的ID',
  comment_content varchar(255) NOT NULL COMMENT '评论内容',
  commenter_user_id mediumint(8) NOT NULL COMMENT '评论者ID',
- comment_time int(13) NOT NULL COMMENT '评论时间',
+ comment_time BIGINT(15) NOT NULL COMMENT '评论时间',
  commenter_ip varchar(15) NOT NULL COMMENT '评论时的IP地址',
  delete_flag bit NOT NULL DEFAULT 0 COMMENT '删除标志',
  PRIMARY KEY (comment_id)
